+++ packages/core/theme/components/a/index.js
@@ -1,6 +1,12 @@
+import { css } from "styled-components";
+
 export const theme = {
   colors: {
     color: "link",
     colorHover: "linkHover"
-  }
+  },
+  fontWeights: {
+    fontWeight: "600"
+  },
+  styles: css``
 };